Refactor in order to deprecate telemetry.
