+++ test/middleware/opentelemetry_tesla_middleware_test.exs
@@ -1,5 +1,270 @@
 defmodule Tesla.Middleware.OpenTelemetryTest do
   use ExUnit.Case
+  require Record
+
+  for {name, spec} <- Record.extract_all(from_lib: "opentelemetry/include/otel_span.hrl") do
+    Record.defrecord(name, spec)
+  end
+
+  for {name, spec} <- Record.extract_all(from_lib: "opentelemetry_api/include/opentelemetry.hrl") do
+    Record.defrecord(name, spec)
+  end
+
+  setup do
+    bypass = Bypass.open()
+
+    :application.stop(:opentelemetry)
+    :application.set_env(:opentelemetry, :tracer, :otel_tracer_default)
+
+    :application.set_env(:opentelemetry, :processors, [
+      {:otel_batch_processor, %{scheduled_delay_ms: 1, exporter: {:otel_exporter_pid, self()}}}
+    ])
+
+    :application.start(:opentelemetry)
+
+    {:ok, bypass: bypass}
+  end
+
+  test "Records spans for Tesla HTTP client", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client) do
+        Tesla.get(client, "/users/")
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect_once(bypass, "GET", "/users", fn conn ->
+      Plug.Conn.resp(conn, 204, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client)
+
+    assert_receive {:span, span(name: "HTTP GET", attributes: attributes)}
+
+    assert %{
+             "http.method": "GET",
+             "http.url": url,
+             "http.target": "/users/",
+             "http.host": "localhost",
+             "http.scheme": "http",
+             "http.status_code": 204
+           } = :otel_attributes.map(attributes)
+
+    assert url == "http://localhost:#{bypass.port}/users/"
+  end
+
+  test "Marks Span status as :error when HTTP request fails", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client) do
+        Tesla.get(client, "/users/")
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect_once(bypass, "GET", "/users", fn conn ->
+      Plug.Conn.resp(conn, 500, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client)
+
+    assert_receive {:span, span(status: {:status, :error, ""})}
+  end
+
+  test "Marks Span status as :errors when max redirects are exceeded", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client) do
+        Tesla.get(client, "/users/")
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry,
+          {Tesla.Middleware.FollowRedirects, max_redirects: 1}
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect(bypass, "GET", "/users", fn conn ->
+      conn
+      |> Plug.Conn.put_resp_header("Location", "/users/1")
+      |> Plug.Conn.resp(301, "")
+    end)
+
+    Bypass.expect(bypass, "GET", "/users/1", fn conn ->
+      conn
+      |> Plug.Conn.put_resp_header("Location", "/users/2")
+      |> Plug.Conn.resp(301, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client)
+
+    assert_receive {:span, span(status: {:status, :error, ""})}
+  end
+
+  test "Appends query string parameters to http.url attribute", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client, id) do
+        params = [id: id]
+        Tesla.get(client, "/users/:id", opts: [path_params: params])
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry,
+          Tesla.Middleware.PathParams,
+          {Tesla.Middleware.Query, [token: "some-token", array: ["foo", "bar"]]}
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect_once(bypass, "GET", "/users/2", fn conn ->
+      Plug.Conn.resp(conn, 204, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client, "2")
+
+    assert_receive {:span, span(name: "HTTP GET", attributes: attributes)}
+
+    mapped_attributes = :otel_attributes.map(attributes)
+
+    assert mapped_attributes[:"http.url"] ==
+             "http://localhost:#{bypass.port}/users/2?token=some-token&array%5B%5D=foo&array%5B%5D=bar"
+  end
+
+  test "http.url attribute is correct when request doesn't contain query string parameters", %{
+    bypass: bypass
+  } do
+    defmodule TestClient do
+      def get(client, id) do
+        params = [id: id]
+        Tesla.get(client, "/users/:id", opts: [path_params: params])
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry,
+          Tesla.Middleware.PathParams,
+          {Tesla.Middleware.Query, []}
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect_once(bypass, "GET", "/users/2", fn conn ->
+      Plug.Conn.resp(conn, 204, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client, "2")
+
+    assert_receive {:span, span(name: "HTTP GET", attributes: attributes)}
+
+    mapped_attributes = :otel_attributes.map(attributes)
+
+    assert mapped_attributes[:"http.url"] ==
+             "http://localhost:#{bypass.port}/users/2"
+  end
+
+  test "Handles url path arguments correctly", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client, id) do
+        params = [id: id]
+        Tesla.get(client, "/users/:id", opts: [path_params: params])
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry,
+          Tesla.Middleware.PathParams,
+          {Tesla.Middleware.Query, [token: "some-token"]}
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    Bypass.expect_once(bypass, "GET", "/users/2", fn conn ->
+      Plug.Conn.resp(conn, 204, "")
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client, "2")
+
+    assert_receive {:span, span(name: "HTTP GET", attributes: attributes)}
+    assert %{"http.target": "/users/2"} = :otel_attributes.map(attributes)
+  end
+
+  test "Records http.response_content_length param into the span", %{bypass: bypass} do
+    defmodule TestClient do
+      def get(client, id) do
+        params = [id: id]
+        Tesla.get(client, "/users/:id", opts: [path_params: params])
+      end
+
+      def client(url) do
+        middleware = [
+          {Tesla.Middleware.BaseUrl, url},
+          Tesla.Middleware.OpenTelemetry,
+          Tesla.Middleware.PathParams,
+          {Tesla.Middleware.Query, [token: "some-token"]}
+        ]
+
+        Tesla.client(middleware)
+      end
+    end
+
+    response = "HELLO 👋"
+
+    Bypass.expect_once(bypass, "GET", "/users/2", fn conn ->
+      Plug.Conn.resp(conn, 200, response)
+    end)
+
+    client = TestClient.client(endpoint_url(bypass.port))
+
+    TestClient.get(client, "2")
+
+    assert_receive {:span, span(name: "HTTP GET", attributes: attributes)}
+
+    mapped_attributes = :otel_attributes.map(attributes)
+
+    {response_size, _} = Integer.parse(mapped_attributes[:"http.response_content_length"])
+    assert response_size == byte_size(response)
+  end
 
   test "Injects distributed tracing headers" do
     OpentelemetryTelemetry.start_telemetry_span(
@@ -23,4 +288,6 @@ defmodule Tesla.Middleware.OpenTelemetryTest do
 
     assert is_binary(traceparent)
   end
+
+  defp endpoint_url(port), do: "http://localhost:#{port}/"
 end
